more work on build player mechanic

+++ FightingGame/Characters/Enemy.cs
@@ -4,18 +4,18 @@ public class Enemy : ICharacter
 {
     public string Name => throw new NotImplementedException();
 
-    public byte Health => throw new NotImplementedException();
+    public int Health => throw new NotImplementedException();
 
-    public byte Attack => throw new NotImplementedException();
+    public int Attack => throw new NotImplementedException();
 
-    public byte Defense => throw new NotImplementedException();
+    public int Defense => throw new NotImplementedException();
 
     public bool IsAlive()
     {
         throw new NotImplementedException();
     }
 
-    public void ReduceHealth(byte attack)
+    public void ReduceHealth(int attack)
     {
         throw new NotImplementedException();
     }
